chore: infrastructure 모듈 persistence 패키지 하위 구조 변경
