+++ Infrastructure-Module/src/main/kotlin/com/bamyanggang/infrastructuremodule/persistence/strongpoint/jpa/entity/StrongPointJpaEntity.java
@@ -1,4 +1,4 @@
-package com.bamyanggang.infrastructuremodule.persistence.domain.strongpoint.jpa.entity;
+package com.bamyanggang.infrastructuremodule.persistence.strongpoint.jpa.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
