+++ Infrastructure-Module/src/main/kotlin/com/bamyanggang/infrastructuremodule/persistence/tag/jpa/entity/TagJpaEntity.java
@@ -1,4 +1,4 @@
-package com.bamyanggang.infrastructuremodule.persistence.domain.experience.jpa.entity;
+package com.bamyanggang.infrastructuremodule.persistence.tag.jpa.entity;
 
 import static jakarta.persistence.FetchType.*;
 
@@ -18,19 +18,15 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Table(name = "experience_content")
-public class ExperienceContentJpaEntity {
+@Table(name = "tag")
+public class TagJpaEntity {
     @Id
-    @Column(name = "experience_content_id")
+    @Column(name = "tag_id")
     private UUID id;
 
-    @Column(columnDefinition = "TEXT")
-    private String question;
-
-    @Column(columnDefinition = "TEXT")
-    private String answer;
+    private String name;
 
     @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "experience_id")
-    private ExperienceJpaEntity experience;
+    @JoinColumn(name = "parent_tag_id")
+    private TagJpaEntity parentTag;
 }
